deepfake-demo-batch/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
mediapipe
